requirements.txt
 APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
install==1.3.5
numpy==1.22.4
pandas==1.4.2
python-dateutil==2.8.2
python-telegram-bot==13.12
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
tornado==6.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
